32535 Completion of the people profile page to show "hours logged this week" in graphical form
hours circle fix to support devimal parts of hours, delete hours also updates the circle now

+++ PSMasterMind/app/modules/people/profile.js
@@ -688,7 +688,7 @@ function( $scope, $state, $stateParams, $filter, Resources, People, AssignmentSe
 		}
 		
 		$scope.totalWeekHours = 0;
-		_.each(profileWeekHours, function(element) { $scope.totalWeekHours += parseInt(element.hours); });
+		_.each(profileWeekHours, function(element) { $scope.totalWeekHours += parseFloat(element.hours); });
 		
 		$scope.initPercentageCircle();
